Move page class registration logic to the HydeKernel
