+++ packages/framework/src/Foundation/FileCollection.php
@@ -16,7 +16,6 @@ use Hyde\Pages\MarkdownPost;
 use Hyde\Support\Filesystem\MediaFile;
 use Hyde\Support\Filesystem\ProjectFile;
 use Hyde\Support\Filesystem\SourceFile;
-use function is_subclass_of;
 
 /**
  * The FileCollection contains all the discovered source and media files,
@@ -37,8 +36,6 @@ use function is_subclass_of;
  */
 final class FileCollection extends BaseFoundationCollection
 {
-    protected array $pageClasses = [];
-
     /**
      * @param  class-string<\Hyde\Pages\Concerns\HydePage>|null  $pageClass
      * @return \Hyde\Foundation\FileCollection<\Hyde\Support\Filesystem\SourceFile>
@@ -69,32 +66,6 @@ final class FileCollection extends BaseFoundationCollection
         return $this->where(fn (ProjectFile $file): bool => $file instanceof MediaFile);
     }
 
-    public function registerPageClass(string $pageClass): self
-    {
-        if (! is_subclass_of($pageClass, HydePage::class)) {
-            throw new \InvalidArgumentException('The specified class must be a subclass of HydePage.');
-        }
-
-        if (! in_array($pageClass, $this->pageClasses, true)) {
-            $this->pageClasses[] = $pageClass;
-        }
-
-        return $this;
-    }
-
-    public function getRegisteredPageClasses(): array
-    {
-        return $this->pageClasses;
-    }
-
-    public function rerunDiscovery(): self
-    {
-        $this->items = [];
-        $this->runDiscovery();
-
-        return $this;
-    }
-
     protected function runDiscovery(): self
     {
         if (Features::hasHtmlPages()) {
@@ -117,7 +88,7 @@ final class FileCollection extends BaseFoundationCollection
             $this->discoverFilesFor(DocumentationPage::class);
         }
 
-        foreach ($this->pageClasses as $pageClass) {
+        foreach ($this->kernel->getRegisteredPageClasses() as $pageClass) {
             $this->discoverFilesFor($pageClass);
         }
